bug/warning fixes.

+++ src/rom_handler.cpp
@@ -638,7 +638,7 @@ void MFEMROMHandler::NonlinearSolve(Operator &oper, BlockVector* U, Solver *prec
    printf("Solve ROM.\n");
    reduced_sol = new BlockVector(rom_block_offsets);
    bool use_restart = config.GetOption<bool>("solver/use_restart", false);
-   double amp = config.GetOption<int>("solver/initial_random_perturbation", 1.0e-5);
+   double amp = config.GetOption<double>("solver/initial_random_perturbation", 1.0e-5);
    if (use_restart)
       ProjectGlobalToDomainBasis(U, reduced_sol);
    else
